Added RDF Conversion lambda
The RDF Conversion lambda converts structured source files to RDF. Files
are read from a folder or S3 bucket and resulting RDF files are written
back to another folder or S3 bucket with the same path, but different
extension.
The conversion applies RML mappings to the input files and uses Carml as
mapping engine.
The current implementation of the conversion lambda is based on Quarkus
3.5.x.
Supported input files are JSONL files and mappings are hard-coded to
provided files. In a future release, this will be changed to read
mapping files and instructions (pre-processing, ...) from a
configuration file.